Add edit button and fix module info update
